requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
supabase>=2.0.0
gotrue